updated the program
Beware of this bulk commit!
Now use forward_list, profiling showed list is not so efficient.
Avoid memory leaks.
Do the timing in the code directly for comparable results.
For profiling the input can be read from a file. If no second file is
handed over to the executable standard in will be read.
Timing with "alternative" data has to be done in a for loop.